+++ src/utils/send-email.ts
@@ -84,7 +84,7 @@ const sendEmail = async (
 export const sendReviewMail = (
   onboardingResponses: OnboardingResponses,
   email: string
-) => {
+) =>
   sendEmail(
     [
       {
@@ -215,19 +215,17 @@ ${
     : ""
 }`
   )
-}
 
 export const sendConfirmMail = (
   recipients: {
     address: string
   }[],
-  firstName: string,
-  lastName: string,
+  fullName: string,
   url: string
-) => {
+) =>
   sendEmail(
     recipients,
-    `Demande d'onboarding de ${firstName} ${lastName}`,
+    `Demande d'onboarding de ${fullName}`,
     `Une demande d'onboarding a été effectuée sur Secrétariat.
 
 En tant qu'administrateur, veuillez en effectuer la revue en suivant le lien :
@@ -246,10 +244,9 @@ ${url}`,
   </a>
 </div>`
   )
-}
 
-export const sendRequestMail = (email: string, url: string) => {
-  return sendEmail(
+export const sendRequestMail = (email: string, url: string) =>
+  sendEmail(
     [
       {
         address: email,
@@ -270,4 +267,3 @@ export const sendRequestMail = (email: string, url: string) => {
       </a>
     </div>`
   )
-}